+++ OngProject/Core/Business/UsersBusiness.cs
@@ -36,7 +36,7 @@ namespace OngProject.Core.Business
             _unitOfWork = unitOfWork;
             _context = context;
             _config = config;
-         }
+        }
 
         private readonly IMapper _mapper;
         public UsersBusiness(IUnitOfWork unitOfWork, OngDbContext context, IMapper mapper)
@@ -60,7 +60,7 @@ namespace OngProject.Core.Business
             {
                 _listAux.Add(new ViewUserDTO(user));
             }
-            return _listAux; 
+            return _listAux;
         }
 
         public Task<User> GetById(int id)
@@ -107,31 +107,5 @@ namespace OngProject.Core.Business
         {
             throw new NotImplementedException();
         }
-
-        public string GetToken(User user)
-        {
-            var tokenHandler = new JwtSecurityTokenHandler();
-
-            var key = Encoding.ASCII.GetBytes(_config["Jwt:Key"]);
-
-            var authClaims = new List<Claim> 
-            {
-                new Claim(ClaimTypes.NameIdentifier, user.Id.ToString()),
-                new Claim(ClaimTypes.Email, user.Email),
-                new Claim(ClaimTypes.Role, user.Role?.Name),
-            };
-
-            var authSigningKey = new SymmetricSecurityKey(key);
-
-            var token = new JwtSecurityToken
-            (
-                expires: DateTime.Now.AddHours(1),
-                claims: authClaims,
-                signingCredentials: new SigningCredentials(authSigningKey, SecurityAlgorithms.HmacSha256)
-            );
-            return tokenHandler.WriteToken(token);
-        }
-
-        
     }
 }
